moved passport configuration into passportConfig.js
