--- backend/config/passport.js
@@ -1,1 +1,0 @@
-//todo: move all passport config in index.js to here
